+++ crates/aranya-fast-channels/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0](https://github.com/aranya-project/aranya-core/compare/aranya-fast-channels-v0.12.0...aranya-fast-channels-v0.13.0) - 2025-09-18
+
+### Other
+
+- generate `ChannelId`s internally ([#410](https://github.com/aranya-project/aranya-core/pull/410))
+
 ## [0.12.0](https://github.com/aranya-project/aranya-core/compare/aranya-fast-channels-v0.11.0...aranya-fast-channels-v0.12.0) - 2025-09-17
 
 ### Other